# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraPoints",
        "description": "LaraPoints REST API skeleton documentation",
        "contact": {
            "email": "ihamzehald@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://larapoints.loc/api",
            "description": "LaraPoints host"
        }
    ],
    "paths": {
        "/user/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the authenticated User",
                "description": "Get the authenticated User",
                "operationId": "UserMe",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerJWTAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "schema": "User",
                        "description": "User ID",
                        "type": "string"
                    },
                    "name": {
                        "schema": "User",
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "schema": "User",
                        "description": "User email",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "schema": "User",
                        "description": "Email verified at",
                        "type": "string"
                    },
                    "created_at": {
                        "schema": "User",
                        "description": "Created at",
                        "type": "string"
                    },
                    "updated_at": {
                        "schema": "User",
                        "description": "Updated at",
                        "type": "string"
                    },
                    "api_token": {
                        "schema": "User",
                        "description": "Api token (used for token auth)",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearerJWTAuth": {
                "type": "http",
                "description": "Oauth2 security",
                "name": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "oauth/token",
                        "scopes": []
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}